Move `core.config.load_object` function into `util.import_`
