+++ src/pydiverse/pipedag/core/config.py
@@ -16,7 +16,6 @@ import yaml
 from box import Box
 
 from pydiverse.pipedag.util.deep_merge import deep_merge
-from pydiverse.pipedag.util.import_ import import_object
 
 if TYPE_CHECKING:
     from pydiverse.pipedag.context import ConfigContext
@@ -371,45 +370,6 @@ def expand_variables(
         return str(variables[name])
 
     return re.sub(r"\{[a-zA-Z_]+[a-zA-Z0-9_]*\}", var_sub, string)
-
-
-def load_object(config_dict: dict):
-    """Instantiates an instance of an object given
-
-    The import path (module.Class) should be specified as the "class" value
-    of the dict. The args section of the dict get used as the instance config.
-
-    If the class defines a `_init_conf_` function, it gets called using the
-    config values, otherwise they just get passed to the class initializer.
-
-    >>> # module.Class(argument="value")
-    >>> load_object({
-    >>>     "class": "module.Class",
-    >>>     "args": {
-    >>>         "argument": "value",
-    >>>     },
-    >>> })
-    """
-
-    if "class" not in config_dict:
-        raise RuntimeError(
-            "Attribute 'class' is missing in configuration "
-            "section that supports multiple backends\n"
-            f"config section: {config_dict}"
-        )
-    cls = import_object(config_dict["class"])
-
-    args = config_dict.get("args", {})
-    if not isinstance(args, dict):
-        raise TypeError(
-            f"Invalid type for args section: {type(args)}\n"
-            f"config section: {config_dict}"
-        )
-    try:
-        init_conf = cls._init_conf_
-        return init_conf(args)
-    except AttributeError:
-        return cls(**args)
 
 
 # Nested Dictionary Utilities
